# frontend/react/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Upstream API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/learning/paths/{path_id}": {
      "get": {
        "tags": [
          "workspace"
        ],
        "summary": "Path Progress",
        "description": "Return a mixed path's course completion and independent selection.",
        "operationId": "path_progress_learning_paths__path_id__get",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathProgress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/paths": {
      "get": {
        "tags": [
          "paths"
        ],
        "summary": "List Paths",
        "description": "List all learning paths.\n\nArgs:\n    _current_user: Authenticated username.\n\nReturns:\n    list[dict]: Path list.",
        "operationId": "list_paths_paths_get",
        "parameters": [
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PathListItem"
                  },
                  "title": "Response List Paths Paths Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "paths"
        ],
        "summary": "Add Path",
        "description": "Create a learning path (any authenticated user).\n\nArgs:\n    payload: Path payload.\n    current_user: Authenticated username.\n\nReturns:\n    dict: Created path.",
        "operationId": "add_path_paths_post",
        "parameters": [
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paths/reviews/summary": {
      "get": {
        "tags": [
          "paths"
        ],
        "summary": "Path Review Summaries",
        "description": "Return average rating + count for each path id.",
        "operationId": "path_review_summaries_paths_reviews_summary_get",
        "parameters": [
          {
            "name": "path_ids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer"
              },
              "description": "Path IDs to summarize",
              "default": [],
              "title": "Path Ids"
            },
            "description": "Path IDs to summarize"
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PathReviewSummaryItem"
                  },
                  "title": "Response Path Review Summaries Paths Reviews Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paths/{path_id}/select": {
      "post": {
        "tags": [
          "paths"
        ],
        "summary": "Select Path",
        "description": "Add a path to the current user's selections.\n\nArgs:\n    path_id: Path ID.\n    current_user: Authenticated username.\n\nReturns:\n    dict: Selection result.",
        "operationId": "select_path_paths__path_id__select_post",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectPathResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paths/{path_id}/unselect": {
      "post": {
        "tags": [
          "paths"
        ],
        "summary": "Unselect Path",
        "description": "Remove a path from the current user's selections.\n\nArgs:\n    path_id: Path ID.\n    current_user: Authenticated username.\n\nReturns:\n    dict: Removal result.",
        "operationId": "unselect_path_paths__path_id__unselect_post",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnselectPathResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paths/{path_id}/status": {
      "post": {
        "tags": [
          "paths"
        ],
        "summary": "Set Path Status",
        "description": "Update the status for a selected path.\n\nArgs:\n    path_id: Path ID.\n    payload: Status payload.\n    current_user: Authenticated username.\n\nReturns:\n    dict: Update result.",
        "operationId": "set_path_status_paths__path_id__status_post",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paths/selected/list": {
      "get": {
        "tags": [
          "paths"
        ],
        "summary": "List Selected Paths",
        "description": "List paths selected by the current user.\n\nArgs:\n    current_user: Authenticated username.\n\nReturns:\n    list[dict]: Selected paths.",
        "operationId": "list_selected_paths_paths_selected_list_get",
        "parameters": [
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectedPathItem"
                  },
                  "title": "Response List Selected Paths Paths Selected List Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paths/{path_id}": {
      "get": {
        "tags": [
          "paths"
        ],
        "summary": "Get Path",
        "description": "Get a learning path by ID.\n\nArgs:\n    path_id: Path ID.\n    _current_user: Authenticated username.\n\nReturns:\n    dict: Path payload.",
        "operationId": "get_path_paths__path_id__get",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "paths"
        ],
        "summary": "Remove Path",
        "description": "Delete a learning path (owner/admin only).\n\nArgs:\n    path_id: Path ID.\n    current_user: Authenticated username.\n\nReturns:\n    dict: Delete result.",
        "operationId": "remove_path_paths__path_id__delete",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePathResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "paths"
        ],
        "summary": "Edit Path",
        "description": "Update a learning path (owner/admin only).\n\nArgs:\n    path_id: Path ID.\n    payload: Path updates.\n    current_user: Authenticated username.\n\nReturns:\n    dict: Updated path.",
        "operationId": "edit_path_paths__path_id__put",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paths/{path_id}/reviews": {
      "get": {
        "tags": [
          "paths"
        ],
        "summary": "List Path Reviews",
        "description": "List reviews for a path.",
        "operationId": "list_path_reviews_paths__path_id__reviews_get",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PathReviewPayload"
                  },
                  "title": "Response List Path Reviews Paths  Path Id  Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "paths"
        ],
        "summary": "Create Path Review",
        "description": "Create a review for a path (any authenticated user).",
        "operationId": "create_path_review_paths__path_id__reviews_post",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathReviewCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PathReviewPayload"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/paths/{path_id}/reviews/{review_id}": {
      "delete": {
        "tags": [
          "paths"
        ],
        "summary": "Delete Path Review",
        "description": "Delete a path review (owner/admin only).",
        "operationId": "delete_path_review_paths__path_id__reviews__review_id__delete",
        "parameters": [
          {
            "name": "path_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Path Id"
            }
          },
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Review Id"
            }
          },
          {
            "name": "x-session-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Session-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletePathReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PathListItem": {
        "properties": {
          "recommendation_note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Note"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "course_count": {
            "type": "integer",
            "title": "Course Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "created_by"
        ],
        "title": "PathListItem",
        "description": "Path list item payload."
      },
      "PathProgress": {
        "properties": {
          "selected": {
            "type": "boolean",
            "title": "Selected"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseProgress"
            },
            "type": "array",
            "title": "Courses"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "selected",
          "courses",
          "completed",
          "total"
        ],
        "title": "PathProgress",
        "description": "Selection state is independent of course completion."
      },
...
      "PathReviewSummaryItem": {
        "properties": {
          "path_id": {
            "type": "integer",
            "title": "Path Id"
          },
          "avg_rating": {
            "type": "number",
            "title": "Avg Rating"
          },
          "review_count": {
            "type": "integer",
            "title": "Review Count"
          }
        },
        "type": "object",
        "required": [
          "path_id",
          "avg_rating",
          "review_count"
        ],
        "title": "PathReviewSummaryItem",
        "description": "Review summary for a path."
      },
...
    }
  }
}